_sphinx удалён, необходимые функции перенесены в flyout_menu
